+++ optaplanner-core/src/main/java/org/optaplanner/core/impl/score/stream/drools/tri/DroolsTriAccumulateFunctionBridge.java
@@ -16,20 +16,16 @@
 
 package org.optaplanner.core.impl.score.stream.drools.tri;
 
-import java.io.ObjectInput;
-import java.io.ObjectOutput;
-import java.util.Map;
 import java.util.function.Function;
 import java.util.function.Supplier;
 
-import org.kie.api.runtime.rule.AccumulateFunction;
 import org.optaplanner.core.api.function.QuadFunction;
 import org.optaplanner.core.api.score.stream.tri.TriConstraintCollector;
-import org.optaplanner.core.impl.score.stream.drools.common.DroolsAccumulateContext;
+import org.optaplanner.core.impl.score.stream.drools.common.DroolsAbstractAccumulateFunctionBridge;
 import org.optaplanner.core.impl.score.stream.drools.common.TriTuple;
 
 final class DroolsTriAccumulateFunctionBridge<A, B, C, ResultContainer_, NewA>
-        implements AccumulateFunction<DroolsAccumulateContext<ResultContainer_>> {
+        extends DroolsAbstractAccumulateFunctionBridge<ResultContainer_, TriTuple<A, B, C>, NewA> {
 
     private final Supplier<ResultContainer_> supplier;
     private final QuadFunction<ResultContainer_, A, B, C, Runnable> accumulator;
@@ -46,58 +42,17 @@ final class DroolsTriAccumulateFunctionBridge<A, B, C, ResultContainer_, NewA>
     }
 
     @Override
-    public DroolsAccumulateContext<ResultContainer_> createContext() {
-        return new DroolsAccumulateContext<>(supplier.get());
+    protected ResultContainer_ newContainer() {
+        return supplier.get();
     }
 
     @Override
-    public void init(DroolsAccumulateContext<ResultContainer_> context) {
-        context.getUndoMap().clear();
+    protected Runnable accumulate(ResultContainer_ container, TriTuple<A, B, C> tuple) {
+        return accumulator.apply(container, tuple._1, tuple._2, tuple._3);
     }
 
     @Override
-    public void accumulate(DroolsAccumulateContext<ResultContainer_> context, Object value) {
-        Map<Object, Runnable> undoMap = context.getUndoMap();
-        if (undoMap.containsKey(value)) {
-            throw new IllegalStateException("Undo for (" + value +  ") already exists.");
-        }
-        TriTuple<A, B, C> values = (TriTuple<A, B, C>) value;
-        Runnable undo = accumulator.apply(context.getContainer(), values._1, values._2, values._3);
-        undoMap.put(value, undo);
+    protected NewA getResult(ResultContainer_ container_) {
+        return finisher.apply(container_);
     }
-
-    @Override
-    public void reverse(DroolsAccumulateContext<ResultContainer_> context, Object value) {
-        Runnable undo = context.getUndoMap().remove(value);
-        if (undo == null) {
-            throw new IllegalStateException("Undo for (" + value +  ") does not exist.");
-        }
-        undo.run();
-    }
-
-    @Override
-    public Object getResult(DroolsAccumulateContext<ResultContainer_> context) {
-        return finisher.apply(context.getContainer());
-    }
-
-    @Override
-    public boolean supportsReverse() {
-        return true;
-    }
-
-    @Override
-    public Class<?> getResultType() {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public void writeExternal(ObjectOutput out) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public void readExternal(ObjectInput in) {
-        throw new UnsupportedOperationException();
-    }
-
 }
