+++ optaplanner-core/src/main/java/org/optaplanner/core/impl/score/stream/drools/uni/DroolsUniAccumulateFunctionBridge.java
@@ -16,19 +16,15 @@
 
 package org.optaplanner.core.impl.score.stream.drools.uni;
 
-import java.io.ObjectInput;
-import java.io.ObjectOutput;
-import java.util.Map;
 import java.util.function.BiFunction;
 import java.util.function.Function;
 import java.util.function.Supplier;
 
-import org.kie.api.runtime.rule.AccumulateFunction;
 import org.optaplanner.core.api.score.stream.uni.UniConstraintCollector;
-import org.optaplanner.core.impl.score.stream.drools.common.DroolsAccumulateContext;
+import org.optaplanner.core.impl.score.stream.drools.common.DroolsAbstractAccumulateFunctionBridge;
 
 final class DroolsUniAccumulateFunctionBridge<A, ResultContainer_, NewA>
-        implements AccumulateFunction<DroolsAccumulateContext<ResultContainer_>> {
+        extends DroolsAbstractAccumulateFunctionBridge<ResultContainer_, A, NewA> {
 
     private final Supplier<ResultContainer_> supplier;
     private final BiFunction<ResultContainer_, A, Runnable> accumulator;
@@ -45,57 +41,17 @@ final class DroolsUniAccumulateFunctionBridge<A, ResultContainer_, NewA>
     }
 
     @Override
-    public DroolsAccumulateContext<ResultContainer_> createContext() {
-        return new DroolsAccumulateContext<>(supplier.get());
+    protected ResultContainer_ newContainer() {
+        return supplier.get();
     }
 
     @Override
-    public void init(DroolsAccumulateContext<ResultContainer_> context) {
-        context.getUndoMap().clear();
+    protected Runnable accumulate(ResultContainer_ container, A tuple) {
+        return accumulator.apply(container, tuple);
     }
 
     @Override
-    public void accumulate(DroolsAccumulateContext<ResultContainer_> context, Object value) {
-        Map<Object, Runnable> undoMap = context.getUndoMap();
-        if (undoMap.containsKey(value)) {
-            throw new IllegalStateException("Undo for (" + value +  ") already exists.");
-        }
-        Runnable undo = accumulator.apply(context.getContainer(), (A) value);
-        undoMap.put(value, undo);
+    protected NewA getResult(ResultContainer_ container_) {
+        return finisher.apply(container_);
     }
-
-    @Override
-    public void reverse(DroolsAccumulateContext<ResultContainer_> context, Object value) {
-        Runnable undo = context.getUndoMap().remove(value);
-        if (undo == null) {
-            throw new IllegalStateException("Undo for (" + value +  ") does not exist.");
-        }
-        undo.run();
-    }
-
-    @Override
-    public Object getResult(DroolsAccumulateContext<ResultContainer_> context) {
-        return finisher.apply(context.getContainer());
-    }
-
-    @Override
-    public boolean supportsReverse() {
-        return true;
-    }
-
-    @Override
-    public Class<?> getResultType() {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public void writeExternal(ObjectOutput out) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public void readExternal(ObjectInput in) {
-        throw new UnsupportedOperationException();
-    }
-
 }
